Revert most of common_test.go
